Added cartReducer and actions

+++ server/db/models/product.js
@@ -1,21 +1,21 @@
-const Sequelize = require('sequelize');
-const db = require('../db');
+const Sequelize = require("sequelize");
+const db = require("../db");
 
-const Product = db.define('product', {
+const Product = db.define("product", {
   name: {
     type: Sequelize.STRING,
     allowNull: false,
   },
   description: {
     type: Sequelize.TEXT,
-    defaultValue: 'Ssssssnakes',
+    defaultValue: "Ssssssnakes",
   },
   imageUrl: {
     type: Sequelize.TEXT,
-    defaultValue: 'https://freepngimg.com/thumb/categories/1402.png',
+    defaultValue: "https://freepngimg.com/thumb/categories/1402.png",
   },
   price: {
-    type: Sequelize.DECIMAL(10, 2),
+    type: Sequelize.INTEGER,
     allowNull: false,
     validate: {
       min: 0,
